Intermediate Points Bug fixed
Reading out intermediate LRPS is now possible. Changed byte arrays to int array to work around signed byte problem.

+++ src/main/java/HereDecoder/MajorMinorVersion.java
@@ -22,7 +22,7 @@ public class MajorMinorVersion {
         return isValid;
     }
 
-    public int decode(byte[] buff) {
+    public int decode(int[] buff) {
         int readBytes = 1;
         //IntUnTi one Byte
         int major = (buff[0] & 0xF0) >> 4;
